+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 
 # Changelog for [`asfa`](https://github.com/obreitwi/asfa)
 
-## v0.4.2-pre (under development)
+## v0.5.0 (2020-10-22)
 
 * Allow host selection via `$ASFA_HOST` environment variable. Priority for host selection is:
   1. `-H`/`--host` supplied via command line.